platform: extract shared ESP shell into esp_shell.c
ESP32-C3 main.c (527 lines) and ESP32-S3 main.c (259 lines) had
massively duplicated shell code: command dispatch, help, thinking
animation, chat handler, and the main REPL loop.  Worse, C3 had
full line editing (arrows, tab completion, UTF-8 aware cursor)
while S3 only had basic raw byte echo — a silent feature gap.

Extract all shared shell logic into platform/common/espressif/
esp_shell.c with a single public entry point esp_shell_loop().
Both main.c files shrink to ~45 lines each (just app_main with
platform init).  ESP32-S3 gains full line editing for free.

Also unify the ESP log level to ESP_LOG_WARN on both platforms
(S3 previously used ESP_LOG_NONE which masked issues).

Before:  C3=527 lines, S3=259 lines (786 total, ~400 duplicated)
After:   C3=45 lines, S3=45 lines, esp_shell.c=423 lines (513 total)
